+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Support thematic breaks.
+- Support paragraph interrupt.
+
 ## 0.3.0
 
 ### Added